# Conflicts:
#	src/EtheriT.Coker.Web.Public/Program.cs

+++ src/EtheriT.Coker.Application.Shared/Dto/Member/MemberGetAllListDto.cs
@@ -15,7 +15,7 @@ namespace EtheriT.Coker.Application.Shared.Dto.Member
         public string? Address { get; set; }
         public string Email { get; set; }
         public int? Total { get; set; }
-        public int? Level { get; set; }
+        public long? Level { get; set; }
         public DateTime CreationTime { get; set; }
 
     }
